Adapted Reexports from bigstatsr

These are functions that were adapted from the bigstatsr package, modified to suppress certain warnings and errors (e.g., from zero variance in scaling).

CutBySize(m, block.size, nb = ceiling(m/block.size))

big_crossprodSelf_no_warning(
  X,
  fun.scaling = big_scale_no_warning(center = FALSE, scale = FALSE),
  ind.row = bigstatsr::rows_along(X),
  ind.col = bigstatsr::cols_along(X),
  block.size = bigstatsr::block_size(nrow(X)),
  backingfile = tempfile(tmpdir = getOption("FBM.dir"))
)

big_cor_no_warning(
  X,
  ind.row = bigstatsr::rows_along(X),
  ind.col = bigstatsr::cols_along(X),
  block.size = bigstatsr::block_size(nrow(X)),
  backingfile = tempfile(tmpdir = getOption("FBM.dir"))
)

big_scale_no_warning(center = TRUE, scale = TRUE)

Arguments

m

An integer specifying the length of the input to split into intervals.

block.size

An integer specifying the maximum length of each block.

nb

Number of blocks. Default is ceiling(m / block.size).

X

An object of class FBM.

fun.scaling

A function with parameters X, ind.row and ind.col, and that returns a data.frame with $center and $scale for the columns corresponding to ind.col, to scale each of their elements such as followed: $$\frac{X_{i,j} - center_j}{scale_j}.$$ Default doesn't use any scaling. You can also provide your own center and scale by using as_scaling_fun().

ind.row

An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col

An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

backingfile

Path to the file storing the FBM data on disk. An extension ".bk" will be automatically added. Default stores in the temporary directory, which you can change using global option "FBM.dir".

center

A logical value: whether to return means or 0s.

scale

A logical value: whether to return standard deviations or 1s. You can't use scale without using center.

Value

Intervals from the input length.

The self crossproduct of an FBM.

The Pearson correlation matrix from an FBM.

A new function that returns a data.frame with two vectors, center and scale, both of length ind.col.

Functions

CutBySize(): Copied version of the unexported bigstatsr:::CutBySize function.

big_crossprodSelf_no_warning(): Adapted version of bigstatsr::big_crossprodSelf that suppresses warnings and errors related to zero scaling.

big_cor_no_warning(): Adapted version of bigstatsr::big_cor that suppresses warnings and errors.

big_scale_no_warning(): Adapted version of bigstatsr::big_scale that suppresses warnings and errors.

Examples

m = 1000
intervals = CutBySize(m, 100)
intervals
#>       lower upper size
#>  [1,]     1   100  100
#>  [2,]   101   200  100
#>  [3,]   201   300  100
#>  [4,]   301   400  100
#>  [5,]   401   500  100
#>  [6,]   501   600  100
#>  [7,]   601   700  100
#>  [8,]   701   800  100
#>  [9,]   801   900  100
#> [10,]   901  1000  100
mat <- matrix(rnorm(400), 20, 20)
X = bigstatsr::FBM(20, 20, init = mat)
crossp <- big_crossprodSelf_no_warning(X)
crossp_base <- crossprod(mat)
all.equal(crossp[], crossp_base)
#> [1] TRUE
mat <- matrix(rnorm(400), 20, 20)
X = bigstatsr::FBM(20, 20, init = mat)
cor <- big_cor_no_warning(X)
cor_base <- cor(mat)
all.equal(cor[], cor_base)
#> [1] TRUE
set.seed(123)
mat <- matrix(rnorm(200), 20, 10)
X <- bigstatsr::FBM(20, 10, init = mat)
bs_ns <- big_scale_no_warning(center = TRUE, scale = TRUE)
scale_stats <- bs_ns(X)
scale_stats
#>         center     scale
#> 1   0.14162380 0.9726653
#> 2  -0.05125716 0.8299387
#> 3   0.10648523 0.9573406
#> 4  -0.11991706 0.9731062
#> 5   0.37509474 0.8289955
#> 6  -0.35938048 0.7010640
#> 7  -0.03834920 1.0853571
#> 8  -0.17333628 1.0319202
#> 9   0.16688984 1.0731386
#> 10 -0.13355787 0.9109602
bigstatsr::big_apply(X, function(X, ind) {
  X.sub <- X[, ind, drop = FALSE]
  X.sub <- t((t(X.sub) - scale_stats$center[ind]) / scale_stats$scale[ind])
  X[, ind] <- X.sub
  NULL
}, block.size = 3)
#> [[1]]
#> NULL
#> 
#> [[2]]
#> NULL
#> 
#> [[3]]
#> NULL
#> 
#> [[4]]
#> NULL
#> 
scaled_mat <- scale(mat)
all.equal(X[], scaled_mat, check.attributes = FALSE)
#> [1] TRUE
X[1:5, 1:5]
#>             [,1]       [,2]       [,3]       [,4]        [,5]
#> [1,] -0.72183046 -1.2248695 -0.8368936  0.5133628 -0.44551577
#> [2,] -0.38224997 -0.2008796 -0.3284124 -0.3929750  0.01228676
#> [3,]  1.45690865 -1.1744811 -1.4330131 -0.2191850 -0.89958849
#> [4,] -0.07311396 -0.8164869  2.1543752 -0.9234946  0.32482906
#> [5,] -0.01268275 -0.6913548  1.1505590 -0.9781812 -0.71843735
scaled_mat[1:5, 1:5]
#>             [,1]       [,2]       [,3]       [,4]        [,5]
#> [1,] -0.72183046 -1.2248695 -0.8368936  0.5133628 -0.44551577
#> [2,] -0.38224997 -0.2008796 -0.3284124 -0.3929750  0.01228676
#> [3,]  1.45690865 -1.1744811 -1.4330131 -0.2191850 -0.89958849
#> [4,] -0.07311396 -0.8164869  2.1543752 -0.9234946  0.32482906
#> [5,] -0.01268275 -0.6913548  1.1505590 -0.9781812 -0.71843735