anglemania_object - Class for Storing and Processing Gene Expression Data

The anglemania_object class is designed to construct the correct input for the anglemania function from a Seurat object and store the results of the analysis. It encapsulates the data and metadata required for processing gene expression data across multiple datasets and batches.

Value

An object of class 'anglemania_object'

Slots

matrix_list

A list of FBM objects containing the gene expression matrices for each batch.

dataset_key

A character string indicating the key used to denote the dataset in the metadata.

batch_key

A character string indicating the key used to denote the batch in the metadata.

data_info

A data frame summarizing the number of samples per dataset and their weights.

weights

A numeric vector of weights for each dataset based on the number of samples.

list_stats

A list containing statistical measures computed across the datasets.

intersect_genes

A character vector of genes that are expressed in at least the specified number of cells across all batches.

min_cells_per_gene

A numeric value indicating the minimum number of cells in which a gene must be expressed to be included in the analysis.

integration_genes

A list containing information about integration genes and their statistics.

See also

Examples

se <- SeuratObject::pbmc_small
se[[]]$Dataset <- rep(c("A", "B"), each = ncol(se) / 2)
angl <- create_anglemania_object(
  se,
  dataset_key = "Dataset",
  batch_key = "groups",
  min_cells_per_gene = 1
)
#> Using dataset_key: Dataset
#> Extracting count matrices...
#> Filtering each batch to at least 1 cells per gene...
#> Using the intersection of filtered genes from all batches...
#> Number of genes in intersected set: 156
#>   |                                                  | 0 % elapsed=00s     |=========================                         | 50% elapsed=00s, remaining~00s  |==================================================| 100% elapsed=00s, remaining~00s
angl
#> anglemania_object
#> --------------
#> Dataset key: Dataset 
#> Batch key: groups 
#> Number of datasets: 2 
#> Datasets: c("A", "B") 
#> Total number of batches: 4 
#> Batches (showing first 5):
#> c("A:g2", "A:g1", "B:g1", "B:g2") 
#> Number of intersected genes: 156 
#> Intersected genes (showing first 10):
#> CD79B, HLA-DRA, HLA-DQB1, HVCN1, HLA-DMB, LTB, EAF2, FAM96A, CXCR4, NT5C , ...
#> Min cells per gene: 1