Calculate statistical measures from a list of FBMs

Computes the mean, standard deviations, and signal-to-noise ratio (SNR) for each element across a list of FBMs in an anglemania_object.

Usage

get_list_stats(angl)

Arguments

angl

An anglemania_object containing the list of FBMs.

Value

A list containing three matrices: mean_zscore, sds_zscore, and sn_zscore.

See also

Examples

se <- SeuratObject::pbmc_small
angl <- create_anglemania_object(se, batch_key = "groups")
#> No dataset_key specified.
#> Assuming that all samples belong to the same dataset and are separated by batch_key: groups
#> Extracting count matrices...
#> Filtering each batch to at least 1 cells per gene...
#> Using the intersection of filtered genes from all batches...
#> Number of genes in intersected set: 228
#>   |                                                  | 0 % elapsed=00s     |==================================================| 100% elapsed=00s, remaining~00s
angl <- anglemania(angl)
#> Computing angles and transforming to z-scores...
#>   |                                                  | 0 % elapsed=00s     |=========================                         | 50% elapsed=00s, remaining~00s  |==================================================| 100% elapsed=00s, remaining~00s
#> Computing statistics...
#> Weighting matrix_list...
#> Calculating mean...
#> Calculating sds...
#> Filtering features...
list_stats(angl) <- get_list_stats(angl)
#> Weighting matrix_list...
#> Calculating mean...
#> Calculating sds...
str(list_stats(angl))
#> List of 3
#>  $ mean_zscore: num [1:228, 1:228] NA 4.02 4.88 2.85 4.48 ...
#>  $ sds_zscore : num [1:228, 1:228] NA 0.601 0.21 0.461 0.86 ...
#>  $ sn_zscore  : num [1:228, 1:228] NA 6.69 23.19 6.17 5.21 ...