Generate example SingleCellExperiment object

This function generates a SingleCellExperiment object with 2 batches and 2 datasets. The object contains 300 genes and 600 cells. The counts matrix is generated using the rpois function with different lambda values for the two batches.

sce_example(seed = 42)

Arguments

seed

The seed for the random number generator. Because this function is only used locally, we allow to set a seed within the function.

Value

A SingleCellExperiment object

Examples

sce <- sce_example()
sce
#> class: SingleCellExperiment 
#> dim: 300 600 
#> metadata(0):
#> assays(1): counts
#> rownames(300): gene1 gene2 ... gene299 gene300
#> rowData names(0):
#> colnames(600): cell1 cell2 ... cell599 cell600
#> colData names(2): batch dataset
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):